update scripts to accept HPG_INPUT=test add -clean option to force clean

+++ bin/run_accel2023.sh
@@ -4,7 +4,7 @@
 # ACCEL2023_BUILD_NUM_THREADS    Number of parallel compile processes. Default: 32
 # export ACC_INPUT=ref   if you wantto run reference isntead of test
 
-export ACC_INPUT=${ACC_INPUT:-test}
+export HPG_INPUT=${HPG_INPUT:-test}
 realpath=`realpath $0`
 thisdir=`dirname $realpath`
 export AOMP_USE_CCACHE=0
@@ -14,15 +14,18 @@ export AOMP_USE_CCACHE=0
 : ${ACCEL2023_SOURCE_DIR:=$AOMP_REPOS_TEST/accel2023-2.0.18}
 : ${ACCEL2023_BUILD_NUM_THREADS:=32}
 
-rm -rf ${ACCEL2023_SOURCE_DIR}
-mkdir -p ${ACCEL2023_SOURCE_DIR}
+if [ "$1" == "-clean" ]; then
+  rm -rf ${ACCEL2023_SOURCE_DIR}
+  mkdir -p ${ACCEL2023_SOURCE_DIR}
+  cd ${ACCEL2023_SOURCE_DIR} || exit 1
+  wget http://roclogin.amd.com/SPEC/accel2023-2.0.18.tar.xz
+  wget http://roclogin.amd.com/SPEC/Accel23-scripts.tar
+  tar xf accel2023-2.0.18.tar.xz
+  ./install.sh -f
+  tar xvf Accel23-scripts.tar
+else
+  cd ${ACCEL2023_SOURCE_DIR} || exit 1
+fi
 
-cd ${ACCEL2023_SOURCE_DIR} || exit 1
-wget http://roclogin.amd.com/SPEC/accel2023-2.0.18.tar.xz
-wget http://roclogin.amd.com/SPEC/Accel23-scripts.tar
-tar xf accel2023-2.0.18.tar.xz
-./install.sh -f
-tar xvf Accel23-scripts.tar
-sed -i s/ref/${ACC_INPUT}/g ./runOne
 ./runOne
 grep ratio= result/*.log
